Updated Self Deployer to use kubectl
