+++ litmus-portal/backend/graphql-server/manifests/subscriber.yml
@@ -26,6 +26,11 @@ spec:
             - name: KEY
               value: #{KEY}
 ---
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: argo
+---
 # This is an auto-generated file. DO NOT EDIT
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
@@ -406,6 +411,7 @@ apiVersion: apps/v1
 kind: Deployment
 metadata:
   name: argo-server
+  namespace: argo
 spec:
   selector:
     matchLabels:
@@ -438,6 +444,7 @@ apiVersion: apps/v1
 kind: Deployment
 metadata:
   name: workflow-controller
+  namespace: argo
 spec:
   selector:
     matchLabels:
