--- litmus-portal/backend/self-deployer/go.mod
@@ -1,17 +1,0 @@
-module github.com/litmuschaos/litmus/litmus-portal/backend/self-deployer
-
-go 1.14
-
-require (
-	github.com/evanphx/json-patch v4.5.0+incompatible // indirect
-	github.com/fsnotify/fsnotify v1.4.9 // indirect
-	github.com/googleapis/gnostic v0.3.1 // indirect
-	github.com/json-iterator/go v1.1.10 // indirect
-	github.com/onsi/gomega v1.10.1 // indirect
-	github.com/pkg/errors v0.8.1 // indirect
-	golang.org/x/text v0.3.3 // indirect
-	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
-	k8s.io/apimachinery v0.18.6
-	k8s.io/client-go v0.18.6
-	k8s.io/utils v0.0.0-20200603063816-c1c6865ac451 // indirect
-)
