--- litmus-portal/backend/self-deployer/pkg/k8s/cluster.go
@@ -1,84 +1,0 @@
-package k8s
-
-import (
-	"context"
-	"flag"
-	"k8s.io/client-go/tools/clientcmd"
-	"log"
-
-	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime/serializer/yaml"
-	"k8s.io/client-go/discovery"
-	memory "k8s.io/client-go/discovery/cached"
-	"k8s.io/client-go/dynamic"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/restmapper"
-)
-
-var decUnstructured = yaml.NewDecodingSerializer(unstructured.UnstructuredJSONScheme)
-
-func CreateDeployment(manifest []byte, kubeconfig *string) (*unstructured.Unstructured, error) {
-	cfg, err := GetKubeConfig(kubeconfig)
-	if err != nil {
-		log.Printf("err: %v\n", err)
-	}
-
-	dc, err := discovery.NewDiscoveryClientForConfig(cfg)
-	if err != nil {
-		log.Printf("err: %v\n", err)
-	}
-	mapper := restmapper.NewDeferredDiscoveryRESTMapper(memory.NewMemCacheClient(dc))
-
-	// Prepare the dynamic client
-	dyn, err := dynamic.NewForConfig(cfg)
-	if err != nil {
-		log.Printf("err: %v\n", err)
-	}
-
-	// Decode YAML manifest into unstructured.Unstructured
-	obj := &unstructured.Unstructured{}
-	_, gvk, err := decUnstructured.Decode([]byte(manifest), nil, obj)
-	if err != nil {
-		log.Printf("err: %v\n", err)
-	}
-
-	// Find GVR
-	mapping, err := mapper.RESTMapping(gvk.GroupKind(), gvk.Version)
-	if err != nil {
-		log.Printf("err: %v\n", err)
-	}
-
-	// Obtain REST interface for the GVR
-	var dr dynamic.ResourceInterface
-	if mapping.Scope.Name() == meta.RESTScopeNameNamespace {
-		// namespaced resources should specify the namespace
-		dr = dyn.Resource(mapping.Resource).Namespace(obj.GetNamespace())
-	} else {
-		// for cluster-wide resources
-		dr = dyn.Resource(mapping.Resource)
-	}
-
-	response, err := dr.Create(context.TODO(), obj, metav1.CreateOptions{})
-	if err != nil {
-		log.Printf("err: %v\n", err)
-	}
-	return response, nil
-}
-
-func GetKubeConfig(kubeconfig *string) (*rest.Config, error) {
-	flag.Parse()
-	// Use in-cluster config if kubeconfig path is specified
-	if *kubeconfig == "" {
-		config, err := rest.InClusterConfig()
-		if err != nil {
-			return config, err
-		}
-	}
-	config, err := clientcmd.BuildConfigFromFlags("", *kubeconfig)
-	if err != nil {
-		return config, err
-	}
-	return config, err
-}
