+++ litmus-portal/backend/tools/self-deployer/go.mod
@@ -1,0 +1,3 @@
+module github.com/litmuschaos/litmus/litmus-portal/backend/tools/self-deployer
+
+go 1.14
