+++ litmus-portal/k8s-manifest.yml
@@ -164,13 +164,14 @@ metadata:
   namespace: litmus
 ---
 apiVersion: rbac.authorization.k8s.io/v1
-kind: RoleBinding
+kind: ClusterRoleBinding
 metadata:
   name: litmus-rb
   namespace: litmus
 subjects:
   - kind: ServiceAccount
     name: litmus-svc-account
+    namespace: litmus
 roleRef:
   kind: ClusterRole
   name: cluster-admin
